pyproject.toml
[project]
name = "kosma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.2.0",
    "pillow>=11.0.0",
    "pytz>=2024.2",
]